# Optimized number to point