@MrDiver 2023-08-05
I would still suggest taking a look at the currently failing tests. There maybe is a way to make it perfectly centered again with some small nudging.