@chopan050 2023-12-10
Converted to draft while I try to figure out what to do with this. I tried some stuff, and it still doesn't pass the tests:
- round the result returned by `n2p` to "fix" the floating point errors
- reverted the calculation of endpoints when there are tips
- reverted almost everything

and it still errors out :sweat_smile: I'll figure out later what to do with this.